% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S06-add_arrow.R
\name{add_arrow}
\alias{add_arrow}
\title{Add Arrow to Diagram}
\usage{
add_arrow(
  x = NULL,
  start = NULL,
  end = NULL,
  col = NULL,
  lwd = NULL,
  length = NULL,
  draw_arrowhead_at = "end",
  check_for_style_guide = T,
  ...
)
}
\arguments{
\item{x}{A vector with 1) the x and y-axis
coordinates, respectively, for the starting point
of the arrow, and 2) the x and y-axis
coordinates, respectively, for the ending point of
the arrow.}

\item{start}{The x and y-axis coordinates for the
starting position of the arrow}

\item{end}{The x and y-axis coordinates for the
ending position of the arrow.}

\item{col}{The color of the arrow}

\item{lwd}{The thickness of the arrow.}

\item{draw_arrowhead_at}{The location at which to draw the
arrow head, either...
\itemize{
  \item 'end';
  \item 'start';
  \item 'both'.
}}

\item{check_for_style_guide}{Logical; if \code{TRUE},
checks if a list called \code{style_guide}
exists in the global environment; if so,
default options for the arrow
are based on the contents of this list.}

\item{...}{Additional arguments for the
\code{\link[graphics]{arrows}} function.}
}
\description{
Adds an arrow to an existing diagram (see
\code{\link{create_base_figure}}.
}
\examples{
# Create diagram
create_base_figure( guidelines = F )
# Add line
add_arrow( c( .25, .25, .75, .75 ), lwd = 3, lty = 2 )

}
