% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-add_line.R
\name{add_line}
\alias{add_line}
\title{Add Line to Diagram}
\usage{
add_line(
  x = NULL,
  start = NULL,
  end = NULL,
  col = NULL,
  lwd = NULL,
  check_for_style_guide = T,
  ...
)
}
\arguments{
\item{x}{A vector with 1) the x and y-axis
coordinates, respectively, for the starting point
of the line, and 2) the x and y-axis
coordinates, respectively, for the ending point of
the line.}

\item{start}{The x and y-axis coordinates for the
starting position of the line.}

\item{end}{The x and y-axis coordinates for the
ending position of the line.}

\item{col}{The color of the line.}

\item{lwd}{The thickness of the line.}

\item{check_for_style_guide}{Logical; if \code{TRUE},
checks if a list called \code{style_guide}
exists in the global environment; if so,
default options for the line
are based on the contents of this list.}

\item{...}{Additional arguments for the
\code{\link[graphics]{segments}} function.}
}
\description{
Adds a line (e.g., to connect nodes)
to an existing diagram (see
\code{\link{create_base_figure}}.
}
\examples{
# Create diagram
create_base_figure( guidelines = F )
# Add line
add_line( c( 0, .5, 1, .5 ), lwd = 3, lty = 2 )

# Add nodes
ndA = add_node( node_content( x = .25, y = .75, text = 'A' ) )
ndB = add_node( node_content( x = .75, y = .25, text = 'B' ) )

# Link nodes
add_line( start = ndA$bottom, end = ndB$top )

}
