% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04-add_node.R
\name{add_node}
\alias{add_node}
\title{Add a Node to a Diagram}
\usage{
add_node(content, check_for_style_guide = T, ...)
}
\arguments{
\item{content}{A list governing the position,
appearance, and content of the node (see the
helper function \code{\link{node_content}}).}

\item{check_for_style_guide}{Logical; if \code{TRUE},
checks if a list called \code{style_guide}
exists in the global environment; if so,
default options for the node generation
are based on the contents of this list.}

\item{...}{Additional arguments for the
\code{\link[graphics]{polygon}} function.}
}
\value{
A named list giving the x and y-axis coordinates for
  the \code{center}, \code{left}, \code{right}, \code{top},
  \code{bottom}, \code{topleft}, \code{bottomleft},
  \code{topright}, and \code{bottomright}.
}
\description{
Adds a node to an existing diagram (see
\code{\link{create_base_figure}}.
}
\examples{
# Create diagram
create_base_figure( guidelines = F )
# Add node
nd = add_node( node_content( x = .5, y = .9, text = 'Hello world' ) )

}
