% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArrow.R
\name{addArrow}
\alias{addArrow}
\title{Add an Arrow Linking Nodes}
\usage{
addArrow(node1, node2, pad = c(0.02, 0), width = c(0.025, 0.04),
  arrow = 0.8, positions = c("right", "left"), ...)
}
\arguments{
\item{node1, node2}{Output from the
\code{\link{addNode}} function.}

\item{pad}{A vector with the adjustment to
1) the left/right and 2) top/bottom
points from which the arrow starts and ends.}

\item{width}{A vector with the width for 1)
the band and 2) the arrow head.}

\item{arrow}{A proportion controlling where the
arrow head begins.}

\item{positions}{A vector giving the points on
\code{node1} and \code{node2}, respectively,
to link via the arrow.}

\item{...}{Additional parameters for the
\code{\link[graphics]{polygon}} function.}
}
\description{
A function that adds an arrow pointing
from a starting node to an end node on
an existing diagram.
}
\examples{
blankPlot()
n1 <- addNode( .25, .25, 'Start' )
n2 <- addNode( .75, .75, 'End' )
addArrow( n1, n2 )

}
