% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNode.R
\name{addNode}
\alias{addNode}
\title{Add Node to an Existing Diagram}
\usage{
addNode(x, y, string, shape = "ellipse", dmn = NULL, fixed = FALSE,
  text.cex = 1, text.adj = NULL, text.col = NULL,
  text.vfont = NULL, text.srt = NULL, ...)
}
\arguments{
\item{x}{The x-axis position for the center of the
node.}

\item{y}{The y-axis position for the center of the
node.}

\item{string}{The label for the node.}

\item{shape}{A character string indicating the
shape of the node, either \code{'ellipse'} or
\code{'rectangle'}.}

\item{dmn}{An optional vector with either 1) the
weights to expand the width and height of the node
based on the dimension of the text, or 2) absolute
values controlling the width and height of the node,
irrespective of text dimensions.}

\item{fixed}{Logical; if \code{TRUE}, indicates
that the input for \code{dmn} represents the
width and height, rather than relative weights.}

\item{text.cex}{Numeric character expansion factor;
multiplied by par("cex") yields the final character
size. NULL and NA are equivalent to 1.0.}

\item{text.adj}{One or two values in [0, 1] which
specify the x (and optionally y) adjustment of
the labels. On most devices values outside that
interval will also work..}

\item{text.col}{the color of the font.}

\item{text.vfont}{\code{NULL} for the current font
family, or a character vector of length 2 for
Hershey vector fonts. The first element of the
vector selects a typeface and the second
element selects a style.}

\item{...}{Additional parameters for the
\code{\link[graphics]{polygon}} function.}
}
\value{
A list giving the Cartesian coordinates for
  different locations of the node.
}
\description{
A function to add a node to an existing diagram.
}
\examples{
blankPlot()
n1 <- addNode( .5, .5, 'Center' )
n2 <- addNode( .25, .25, 'Bottom left', shape = 'rectangle' )

}
