% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-create_base_figure.R
\name{create_base_figure}
\alias{create_base_figure}
\title{Create Base Figure for Diagram}
\usage{
create_base_figure(
  default = NULL,
  w = 6,
  h = 6,
  orientation = "landscape",
  margin = rep(0.25, 4),
  guidelines = T,
  guide_positions = seq(0.1, 0.9, 0.1),
  guide_adjust = 3/4,
  new = T
)
}
\arguments{
\item{default}{Pre-packaged figure sizes, including...
\itemize{
  \item 'US letter' (11 x 8.5 inches);
  \item '4.6' (4.6 x 4.6 inches);
  \item '9.2 x 4.6' (in inches).
}}

\item{w}{Width (in inches) for the x-axis.}

\item{h}{Height (in inches) for the y-axis.}

\item{orientation}{For default specifications, indicates
whether figures should be in \code{landscape} or
\code{portrait} style.}

\item{margin}{A vector with the margins (in inches) for
the bottom, left, top, and right, respectively.}

\item{guidelines}{Logical; if \code{TRUE}, includes
guidelines when generating the figure.}

\item{guide_positions}{Vector giving the guideline
positions (values must be between 0 and 1).}

\item{guide_adjust}{An adjustment controlling the
position on the axes for the guideline numbers.}

\item{new}{Logical; if \code{TRUE} a new plotting
window is generated.}
}
\description{
A function that generates a base figure,
with optional guidelines, for a diagram.
}
\examples{
# Default (6 x 6 inches)
create_base_figure()

# No guidelines
create_base_figure( guidelines = FALSE )

# US letter size
create_base_figure( default = 'US letter' )

# US letter (Portrait)
create_base_figure( default = 'US letter', orientation = 'portrait' )

}
