% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S07-create_style_guide.R
\name{create_style_guide}
\alias{create_style_guide}
\title{Create a Style Guide to Change Global Diagram Defaults}
\usage{
create_style_guide(
  node.shape = "rectangle",
  node.spacing = 0.5,
  node.border_color = "black",
  node.border_width = 2,
  node.node_color = "white",
  node.text_size = 1,
  node.text_color = "black",
  node.alignment = "center",
  line.col = "black",
  line.lwd = 2,
  arrow.col = "black",
  arrow.lwd = 2,
  arrow.length = 0.25
)
}
\arguments{
\item{node.shape}{Specify the default shape for nodes.}

\item{node.spacing}{Specify the default spacing between
text lines in nodes.}

\item{node.border_color}{Specify the default line color
for node borders.}

\item{node.border_width}{Specify the default line
width for node borders.}

\item{node.node_color}{Specify the default fill color
for nodes.}

\item{node.text_size}{Specify the default text size
for nodes.}

\item{node.text_color}{Specify the default text color
for nodes.}

\item{node.alignment}{Specify the default text
alignment for nodes.}

\item{line.col}{Specify the default line
color for lines between nodes.}

\item{line.lwd}{Specify the default line width
for lines between nodes.}

\item{arrow.col}{Specify the default line
color for arrows.}

\item{arrow.lwd}{Specify the default line
width for arrows.}

\item{arrow.length}{Specify the default
length of the arrow heads.}
}
\value{
Creates a named list called \code{style_guide}
  in the global environment.
}
\description{
Creates a list called \code{stye_guide} in the global
environment that governs the defaults for
\code{\link{add_node}}), \code{\link{add_line}}), and
\code{\link{add_arrow}}). Allows widespread changes
to the appearance of nodes, lines, and arrows.
}
\examples{
# Two-panel figure
layout( cbind( 1, 2 ) )

# Standard defaults
create_base_figure( guidelines = F, new = F )
nd = add_node( node_content( text = 'Hello world' ) )
add_line( start = nd$bottom, end = c( .5, .1 ) )

create_style_guide( node.text_size = 2, line.lwd = 3 )
create_base_figure( guidelines = F, new = F )

nd = add_node( node_content( text = 'Hello world' ) )
add_line( start = nd$bottom, end = c( .5, .1 ) )

}
