% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-node_content.R
\name{node_content}
\alias{node_content}
\title{Initialize List for Node Content}
\usage{
node_content(
  x = 0.5,
  y = 0.5,
  text = "",
  dimensions = NULL,
  shape = NULL,
  spacing = NULL,
  border_color = NULL,
  border_width = 2,
  node_color = NULL,
  text_size = NULL,
  text_color = NULL,
  alignment = NULL
)
}
\arguments{
\item{x}{The x-axis position of the node (the center).}

\item{y}{The y-axis position of the node (the top).}

\item{dimensions}{An optional vector giving the width
and height of the node (assuming a relative plot width
and height of 1).}

\item{shape}{The shape of the node, with options for...
\itemize{
  \item 'rectangle'.
}}

\item{spacing}{The relative space between text lines
(defaults to .5 of the text height).}

\item{border_color}{The color of the border for
the node.}

\item{node_color}{The fill color for the node.}

\item{text_size}{The size of the node text.}

\item{text_color}{The color of the text.}

\item{alignment}{The alignment of the text, either...
\itemize{
  \item 'center';
  \item 'left'.
}}

\item{text_content}{Either 1) a character vector with
each line of text, or 2) a named list with the elements...
\itemize{
  \item \code{left}: Left-most text content.
  \item \code{special}: Text content with special
    formatting (e.g., bold text).
  \item \code{right} Right-most text content.
  \item \code{font} The type of formatting.
}}
}
\value{
A list.
}
\description{
Generates a list with all of the elements
used to generate the content of a node.
A helper function for \code{\link{add_node}}.
}
