% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-padding.R
\name{padding}
\alias{padding}
\title{Create Padding for Character Strings}
\usage{
padding(n_spaces)
}
\arguments{
\item{n_spaces}{A vector giving the number of spaces to
use when creating indents.}
}
\value{
A character vector.
}
\description{
Given a vector specifying number of spaces, will
create indents to pass in to the 'left' argument
of the 'node_content' function.
}
\examples{
padding( c( 0, 2, 5 ) )

}
