% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-pd_functions.R
\name{pd_base_figure}
\alias{pd_base_figure}
\title{Create Base Figure}
\usage{
pd_base_figure(
  default = NULL,
  width = 6,
  height = 6,
  orientation = "landscape",
  margin = rep(0.1, 4),
  guidelines = TRUE,
  guide_major = seq(0.1, 0.9, 0.1),
  guide_minor = seq(0.05, 0.95, 0.1),
  guide_adjust = 0.8,
  new = FALSE
)
}
\arguments{
\item{default}{A character string used to specify
pre-packaged figure sizes, including...
\itemize{
  \item 'US letter' or '8.5 x 11' (inches);
  \item '3.54 x 3.54' (inches);
  \item '5 x 5' (inches);
  \item '7.25 x 5' (inches).
}}

\item{width}{A numeric value, the width of the
figure in inches.}

\item{height}{A numeric value, the height of the
figure in inches.}

\item{orientation}{A character string, indicates
whether figures should be in \code{'landscape'} or
\code{'portrait'} style.}

\item{margin}{A numeric vector of 4 values, specifying
in inches the margins for the bottom, left, top, and
right, respectively.}

\item{guidelines}{A logical value, if \code{TRUE} includes
guidelines when generating the figure.}

\item{guide_major}{A numeric vector giving the major guideline
positions (values must be between 0 and 1).}

\item{guide_minor}{A numeric vector giving the minor guideline
positions (values must be between 0 and 1).
If \code{NULL}, no minor guidelines are included.}

\item{guide_adjust}{A numeric value, the adjustment controlling the
position on the axes for the guideline numbers.}

\item{new}{A logical value, if \code{TRUE} a new plotting
window is generated.}
}
\description{
A function that generates a base figure,
with optional guidelines, for a diagram.
}
\examples{
\dontrun{
# Default (6 x 6 inches)
pd_base_figure()

# No guidelines
pd_base_figure( guidelines = FALSE )

# US letter size
pd_base_figure( default = 'US letter' )

# US letter (Portrait)
pd_base_figure( default = 'US letter', orientation = 'portrait' )
}

}
