% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-pd_functions.R
\name{pd_node_dimensions}
\alias{pd_node_dimensions}
\title{Node Dimensions}
\usage{
pd_node_dimensions(xy, wh)
}
\arguments{
\item{xy}{A named vector with the
x and y-axis coordinates for the
center of the node.}

\item{wh}{A named vector with the
width and height of the node.}
}
\value{
A list with the assorted coordinates
for the node.
}
\description{
Function to compute the dimensions
of a node, providing values for the
top, topright, right, bottomright,
bottom, bottomleft, left, and topleft
x and y-axis coordinates.
}
\examples{
pd_node_dimensions( c(x = .5, y = .5), wh = c( w = .1, h = .1 ) )

}
