% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-pd_functions.R
\name{pd_node_template}
\alias{pd_node_template}
\title{Create Template for Nodes}
\usage{
pd_node_template(n)
}
\arguments{
\item{n}{A integer value, the number of nodes
to include in the template code.}
}
\value{
A message to the console window with the
template code.
}
\description{
Function to generate template code for
creating nodes in a path diagram.
}
