% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-legacy_functions.R
\name{replace_with_values}
\alias{replace_with_values}
\title{Replace Placeholders with Values}
\usage{
replace_with_values(
  x,
  values,
  placeholder = function(i) {
     paste0("[[", i, "]]")
 },
  ...
)
}
\arguments{
\item{x}{A character vector}

\item{values}{A vector of values with
which to replace placeholder text.}

\item{placeholder}{A function defining
how to identify placeholder text based
on the index position of the vector
\code{values}.}

\item{...}{Additional parameters for the
\code{placeholder} function.}
}
\description{
Replaces placeholder text in a
character vector with user specified
values.
}
\examples{
# Character vector with placeholder text
# '[[i]]' for the ith element to replace
x <- c( 'Value: [[1]]', 'Value: [[2]]' )
replace_with_values( x, c( 1, 2 ) )

# Custom function to replace different
# placeholder text
placeholder <- function( i ) paste0( '***', LETTERS[i] )
x <- c( 'Value: ***A', 'Value: ***B' )
replace_with_values( x, c( 1, 2 ), placeholder )

}
