% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S11-size_sep.R
\name{sep}
\alias{sep}
\title{Determine Separation Between Nodes}
\usage{
sep(nodes, labels, dimension = "height", desired = NULL)
}
\arguments{
\item{nodes}{...}

\item{labels}{...}

\item{dimension}{...}

\item{desired}{...}
}
\description{
...
}
\examples{
# Forthcoming


}
