% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-pd_functions.R
\name{wh}
\alias{wh}
\title{Vector for width and height}
\usage{
wh(w, h)
}
\arguments{
\item{w}{A numeric value between 0 and 1.}

\item{h}{A numeric value between 0 and 1.}
}
\value{
A named vector.
}
\description{
Function to produce named vector with width
and height to pass to [pathdiagrams::pd_node].
}
