% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-pd_functions.R
\name{xy}
\alias{xy}
\title{Vector for X and Y Coordinates}
\usage{
xy(x, y)
}
\arguments{
\item{x}{A numeric value between 0 and 1.}

\item{y}{A numeric value between 0 and 1.}
}
\value{
A named vector.
}
\description{
Function to produce named vector with x and
y coordinates to pass to [pathdiagrams::pd_node].
}
