% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-add_node_shape.R
\name{add_node_shape}
\alias{add_node_shape}
\title{Add Node Shape}
\usage{
add_node_shape(
  nd,
  shape = "box",
  shape.col = "white",
  shape.lwd = 2,
  shape.border = "black",
  shape.lty = 1,
  shape.x = NA,
  shape.y = NA,
  xpd = NA
)
}
\arguments{
\item{nd}{A named list of vectors with the Cartesian
coordinates for the bottom, left, top, and right of
the internal text box for the node.}

\item{shape}{The default shape for nodes, either 'box',
or 'circle'.}

\item{shape.col}{The fill color for the node.}

\item{shape.lwd}{The line wide for the node border.}

\item{shape.border}{The mode border color
(\code{NA} will suppress the border).}

\item{shape.lty}{The line type for the node.}

\item{shape.x}{The fixed width for the x-axis.}

\item{shape.y}{The fixed height for the y-axis.}

\item{xpd}{A logical value or NA. If \code{FALSE}, all
plotting is clipped to the plot region, if
\code{TRUE}, all plotting is clipped to the figure
region, and if NA, all plotting is clipped to the
device region.}
}
\description{
Given a list with the x and y-axis coordinates for the
bottom, left, top, and right of the internal text box,
draws a specified shape on an existing plot.
}
\examples{
# Empty plot
create_base_figure()

# Named list of coordinates
nd = list(
  top = c( .5, .6 ),
  bottom = c( .5, .4 ),
  left = c( .4, .5 ),
  right = c( .6, .5 )
)

# Rectangle
add_node_shape( nd )
# Ellipse
add_node_shape(
  nd, shape = 'circle',
  shape.border = 'blue', shape.lty = 2
)

}
