% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S00-pathdiagrams-package.R
\docType{package}
\name{pathdiagrams-package}
\alias{pathdiagrams}
\alias{pathdiagrams-package}
\title{Tools to Create Path Diagrams}
\description{
The \pkg{pathdiagrams} package provides a set of
functions to streamline creation of path diagrams
and flow charts. A complete list of the
functions available in the package can be obtained
via the command \code{ls(pos = "package:pathdiagrams")}.
A website with function documentation and vignettes on
how to use the package is available at
https://rettopnivek.github.io/pathdiagrams/.
}
\author{
\strong{Maintainer}: Kevin Potter \email{kevin.w.potter@gmail.com}

}
